UGC-4582 | Fix creating tilesheets via the API without a summary (#7)
This broke after the MW 1.39 upgrade, as string typehints were added to
ManualLogEntry which caused errors as the API would attempt to pass
`null` there when no summary parameter was provided. As a fix, have the
API default to an empty string value if this parameter is absent.

As we're here, add basic continuous integration and create an
integration test for this API module.